feat: Many bugs fixes and features like API endpoints and facebook and google login using JavaScript and server side

+++ vagrant/catalog/catalog/views/profile.py
@@ -9,8 +9,8 @@
  * Time: 14:28
  * To change this template use File | Settings | File Templates.
 """
-from flask import Blueprint, render_template, flash, redirect, url_for
-from flask.ext.login import login_required
+from flask import Blueprint, render_template, flash, redirect, url_for, abort
+from flask.ext.login import login_required, current_user
 
 from catalog import db, app
 from catalog.forms import CategoryForm, ItemForm, DeleteForm
@@ -24,7 +24,7 @@ profile = Blueprint('profile', __name__)
 def new_category():
     form = CategoryForm()
     if form.validate_on_submit():
-        new_category = Category(name=form.name.data)
+        new_category = Category(name=form.name.data, user=current_user)
         db.session.add(new_category)
         db.session.commit()
         if app.debug:
@@ -34,7 +34,7 @@ def new_category():
                 )
             )
         flash("Category {} added!".format(
-            (new_category.id, new_category.name))
+            (new_category.id, new_category.name)), 'success'
         )
         return redirect(url_for('home.index'))
     return render_template("profile/category.html", action="Add",
@@ -46,7 +46,12 @@ def new_category():
 @profile.route('/category/<int:category_id>/edit', methods=['GET', 'POST'])
 @login_required
 def edit_category(category_id):
-    category = Category.query.filter_by(id=category_id).one()
+    category = Category.query.filter_by(id=category_id).first_or_404()
+    owner = category.user
+    if current_user.id != owner.id:
+        flash("You are not authorized to edit this category. "
+              "Please create your own category in order to edit.", 'warning')
+        return redirect(url_for('home.index'))
     form = CategoryForm(obj=category)
     if form.validate_on_submit():
         category.name = form.name.data
@@ -59,7 +64,7 @@ def edit_category(category_id):
                 )
             )
         flash("Category {} edited!".format(
-            (category.id, category.name))
+            (category.id, category.name)), 'success'
         )
         return redirect(url_for('home.index'))
     return render_template("profile/category.html", action="Edit",
@@ -70,7 +75,12 @@ def edit_category(category_id):
 @profile.route('/category/<int:category_id>/delete', methods=['GET', 'POST'])
 @login_required
 def delete_category(category_id):
-    category = Category.query.filter_by(id=category_id).one()
+    category = Category.query.filter_by(id=category_id).first_or_404()
+    owner = category.user
+    if current_user.id != owner.id:
+        flash("You are not authorized to delete this category. "
+              "Please create your own category in order to delete.", 'warning')
+        return redirect(url_for('home.index'))
     form = DeleteForm()
     if form.validate_on_submit():
         db.session.delete(category)
@@ -79,7 +89,7 @@ def delete_category(category_id):
             app.logger.debug("Category {} deleted!".format(
                 (category.id, category.name)))
         flash("Category {} deleted!".format(
-            (category.id, category.name)))
+            (category.id, category.name)), 'success')
         return redirect(url_for('home.index'))
     else:
         return render_template('profile/delete.html', form=form,
@@ -88,10 +98,24 @@ def delete_category(category_id):
                                object=category)
 
 
+@profile.route('/category/<int:category_id>', defaults={'page': 1})
+@profile.route('/category/<int:category_id>/page/<int:page>')
+def category_items(category_id, page):
+    categories = Category.query.all()
+    sel_category = Category.query.filter_by(id=category_id).first_or_404()
+    items = Item.query.filter_by(category_id=category_id).paginate(
+        page, app.config['POSTS_PER_PAGE'], False)
+    if not items and page != 1:
+        abort(404)
+    return render_template('profile/category_items.html', pagination=items,
+                           items=items.items, sel_category=sel_category,
+                           categories=categories)
+
+
 @profile.route('/item/<int:item_id>')
 def item_profile(item_id):
-    item = Item.query.filter_by(id=item_id).one()
-    owner = User.query.filter_by(id=item.user_id).one()
+    item = Item.query.filter_by(id=item_id).first_or_404()
+    owner = User.query.filter_by(id=item.user_id).first_or_404()
     return render_template('profile/item_profile.html', item=item, owner=owner)
 
 
@@ -105,9 +129,10 @@ def new_item():
     else:
         form.category.choices = [(0, 'None')]
     if form.validate_on_submit():
-        category = Category.query.filter_by(id=form.category.data).one()
+        category = Category.query.filter_by(id=form.category.data).first_or_404()
         new_item = Item(name=form.name.data, description=form.description.data,
-                        picture=form.picture.data, category=category)
+                        picture=form.picture.data, category=category,
+                        user=current_user)
         db.session.add(new_item)
         db.session.commit()
         if app.debug:
@@ -117,7 +142,7 @@ def new_item():
                 )
             )
         flash("Item {} added!".format(
-            (new_item.id, new_item.name))
+            (new_item.id, new_item.name)), 'success'
         )
         return redirect(url_for('home.index'))
     return render_template("profile/item.html", action="Add",
@@ -129,15 +154,22 @@ def new_item():
 @profile.route('/item/<int:item_id>/edit', methods=['GET', 'POST'])
 @login_required
 def edit_item(item_id):
-    item = Item.query.filter_by(id=item_id).one()
+    item = Item.query.filter_by(id=item_id).first_or_404()
+    owner = item.user
+    if current_user.id != owner.id:
+        flash("You are not authorized to edit this item. "
+              "Please create your own item in order to edit.", 'warning')
+        return redirect(url_for('home.index'))
+
     form = ItemForm(obj=item)
     categories = Category.query.all()
     if categories:
         form.category.choices = [(cat.id, cat.name) for cat in categories]
+        form.category.data = item.category_id
     else:
         form.category.choices = [(0, 'None')]
     if form.validate_on_submit():
-        category = Category.query.filter_by(id=form.category.data).one()
+        category = Category.query.filter_by(id=form.category.data).first_or_404()
         item.name = form.name.data
         item.description = form.description.data
         item.picture = form.picture.data
@@ -151,7 +183,7 @@ def edit_item(item_id):
                 )
             )
         flash("Item {} edited!".format(
-            (item.id, item.name))
+            (item.id, item.name)), 'success'
         )
         return redirect(url_for('home.index'))
     return render_template('profile/item.html', action='Edit',
@@ -163,7 +195,12 @@ def edit_item(item_id):
 @profile.route('/item/<int:item_id>/delete', methods=['GET', 'POST'])
 @login_required
 def delete_item(item_id):
-    item = Item.query.filter_by(id=item_id).one()
+    item = Item.query.filter_by(id=item_id).first_or_404()
+    owner = item.user
+    if current_user.id != owner.id:
+        flash("You are not authorized to delete this item. "
+              "Please create your own item in order to delete.", 'warning')
+        return redirect(url_for('home.index'))
     form = DeleteForm()
     if form.validate_on_submit():
         db.session.delete(item)
@@ -172,10 +209,20 @@ def delete_item(item_id):
             app.logger.debug("Item {} deleted!".format(
                 (item.id, item.name)))
         flash("Category {} deleted!".format(
-            (item.id, item.name)))
+            (item.id, item.name)), 'success')
         return redirect(url_for('home.index'))
     else:
         return render_template('profile/delete.html', form=form,
                                action='item',
                                form_action='profile.delete_item',
                                object=item)
+
+
+@profile.route('/user/<int:user_id>/profile')
+@login_required
+def user_profile(user_id):
+    user = User.query.filter_by(id=user_id).first_or_404()
+    if current_user.id != user.id:
+        flash("You are not authorized to view this user profile.", 'warning')
+        return redirect(url_for('home.index'))
+    return render_template('profile/user_profile.html', user=user)
